Scripts to load data from the holidays gem
The repositories found on Github at holidays/holidays and
holidays/definitions make up a very similar project to ours! They have a
config repo with YAML files and a Ruby gem that includes it as a git
submodule.

I wish we'd found them years ago!

Their biggest improvement over our system is that their YAML files
define rules, such as "Thanksgiving always falls on the fourth Thursday
of November", and so they don't have to manually update their data as
often as we do.

Their biggest downside over our system is that their regions don't have
names associated with them.

In the future we might switch over completely to their gem, or wrap it
in a new set of config for naming regions, or something. But right now
in December 2020 we're just looking for a quick way to add data for 2021
before the clock runs out 😅